// generated/graph.ts
export declare namespace Graph {
// ...
  export interface HrEmployee {
    id?: null | number;
    username?: null | string;
    fullname?: null | string;
    fullname_en?: null | string;
    profile?: null | string;
    phoneNumber?: null | string;
    status?: null | boolean;
    email?: null | string;
    gender?: null | string;
    nationality?: null | string;
    dob?: null | string;
    district?: null | string;
    commune?: null | string;
    education_level?: null | string;
    passport_id?: null | string;
    national_id?: null | string;
    position_level?: null | string;
    position_description?: null | string;
    unit?: null | string;
    department_id?: null | number;
    general_department_id?: null | number;
    contact_city_or_province?: null | string;
    province?: null | string;
    homeNo?: null | string;
    streetNo?: null | string;
    village_or_group?: null | string;
    contact_district?: null | string;
    contact_village?: null | string;
    contact_commune?: null | string;
    officer_id?: null | string;
    office_id?: null | number;
  }
// ...
}